requirements.txt
# Core framework
Django>=6.0,<7.0
python-dotenv>=1.2,<2.0
dj-database-url>=3.1,<4.0

# Serving and static files
whitenoise>=6.12,<7.0
gunicorn>=24.1,<25.0

# AI integration
openai>=2.16,<3.0

# Media and file handling
Pillow>=12.1,<13.0
python-magic>=0.4.27,<0.5
edge-tts>=7.2,<8.0
gTTS>=2.5,<3.0

# Optional production database driver (PostgreSQL)
psycopg2-binary>=2.9,<3.0
